attempted fix

+++ include/datatypes.h
@@ -50,7 +50,7 @@ class IndexReference {
     IndexReference operator+(const int other) const { return IndexReference(sentence, token+ other); };
     
     std::string tostring() const {
-        return std::to_string(sentence) + ":" + std::to_string(token);
+        return std::to_string((unsigned int) sentence) + ":" + std::to_string((unsigned int) token);
     }
 };
